+++ tests/test_10_nichtparasoup/test_imagecrawler/test_knownimagecrawlers/test_knownimagecrawlers.py
@@ -5,9 +5,7 @@ from pkg_resources import Distribution, EntryPoint
 
 from nichtparasoup.imagecrawler import KnownImageCrawlers
 
-from .testdata_imagecrawler_plugin import (
-    BaseLoadableImageCrawlerA, LoadableImageCrawlerA, LoadableImageCrawlerB, StandaloneLoadableImageCrawlerA,
-)
+from . import BaseLoadableImageCrawlerA, LoadableImageCrawlerA, LoadableImageCrawlerB, StandaloneLoadableImageCrawlerA
 
 _TEST_PLUGIN_DIST = Distribution(location=path.abspath('testdata_imagecrawler_plugin'),
                                  project_name='testdata-imagecrawler-plugin')
@@ -187,7 +185,7 @@ class KnownImageCrawlersAddTest(unittest.TestCase):
         kic = KnownImageCrawlers([])
         kic._list = []
         # act
-        kic._add(_TEST_PLUGIN_ENTRY)
+        kic._append(_TEST_PLUGIN_ENTRY)
         # arrange
         self.assertListEqual([(_TEST_PLUGIN_ENTRY.name, _TEST_PLUGIN_CLASS)], kic._list)